+++ src/main/groovy/io/seqera/wave/configuration/PackagesConfig.groovy
@@ -15,34 +15,49 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.seqera.wave.service.packages
 
+package io.seqera.wave.configuration
+
+import java.time.Duration
+import javax.annotation.PostConstruct
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-import io.micronaut.scheduling.TaskExecutors
-import io.micronaut.scheduling.annotation.ExecuteOn
-import io.micronaut.scheduling.annotation.Scheduled
-import jakarta.inject.Inject
+import io.micronaut.context.annotation.Value
+import io.micronaut.core.annotation.Nullable
 import jakarta.inject.Singleton
 
 /**
- * Cron job to fetch conda packages
+ * Conda Fetcher service settings
  *
  * @author Munish Chouhan <munish.chouhan@seqera.io>
  */
-
 @CompileStatic
 @Singleton
 @Slf4j
-class PackageFetcherCronJob {
+class PackagesConfig {
+    /**
+     * Docker image of tool need to be used for conda fetcher
+     */
+    @Value('${wave.package.conda.image.name:continuumio/miniconda3}')
+    String condaImage
 
-    @Inject
-    PackagesService service
+    @Value('${wave.package.timeout:10m}')
+    Duration timeout
 
-    @ExecuteOn(TaskExecutors.SCHEDULED)
-    @Scheduled(initialDelay = '${wave.package.cron.delay:1m}' , fixedDelay = '${wave.package.cron.interval:6h}')
-    void fetch(){
-        service.fetchPackages()
+    @Value('${wave.package.channels:seqera}')
+    List<String> channels
+
+    @Value('${wave.package.k8s.resources.requests.cpu}')
+    @Nullable
+    String requestsCpu
+
+    @Value('${wave.package.k8s.resources.requests.memory}')
+    @Nullable
+    String requestsMemory
+
+    @PostConstruct
+    private void init() {
+        log.debug("Package config: docker image name: ${condaImage}; timeout: ${timeout}; cpus: ${requestsCpu}; mem: ${requestsMemory}")
     }
 }
