+++ src/main/groovy/io/seqera/wave/service/packages/KubeCondaFetcher.groovy
@@ -24,7 +24,7 @@ import java.time.Instant
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import io.micronaut.context.annotation.Requires
-import io.seqera.wave.configuration.PackageConfig
+import io.seqera.wave.configuration.PackagesConfig
 import io.seqera.wave.service.k8s.K8sService
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
@@ -42,7 +42,7 @@ import jakarta.inject.Singleton
 class KubeCondaFetcher extends AbstractCondaFetcher {
 
     @Inject
-    PackageConfig config
+    PackagesConfig config
 
     @Inject
     K8sService k8sService
